Refactor MobileNav into own component
